requirements.txt
Flask>=2.0
pandas>=1.3.0
numpy>=1.20.0
psycopg2-binary>=2.9.0
requests>=2.26.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
tensorflow>=2.10.0
keras>=2.10.0
matplotlib>=3.5.0
seaborn>=0.12.0
statsmodels>=0.13.0
joblib>=1.2.0
